Add & Remove from basket func finished
The whole process of adding an item to basket. The icon basket itself is being rendered once the add to basket button is clicked on the selected item. Then, clicking on the Basket Icon it directs you to basket page, where you could see the subtotal items and the calculated price. Moreover, if you want to delete an item from the basket, once you click on Delete from basket, the sum will be re-calculated and updated, either the subtotal items.